feat: add getuser, listusers, updateuser

+++ internal/core/port/user.go
@@ -11,7 +11,7 @@ type UserRepository interface {
 	GetUserByEmail(ctx context.Context, email string) (*domain.User, error)
 	GetUserByUsername(ctx context.Context, username string) (*domain.User, error)
 	GetUserByID(ctx context.Context, id string) (*domain.User, error)
-	ListUsers(ctx context.Context, page, limit int64) ([]domain.User, error)
+	ListUsers(ctx context.Context, page, limit int) ([]domain.User, error)
 	UpdateUser(ctx context.Context, user *domain.User) (*domain.User, error)
 	DeleteUser(ctx context.Context, id string) error
 }
@@ -19,7 +19,7 @@ type UserRepository interface {
 type UserService interface {
 	Register(ctx context.Context, user *domain.User) (*domain.User, error)
 	GetUser(ctx context.Context, id string) (*domain.User, error)
-	ListUsers(ctx context.Context, page, limit int64) ([]domain.User, error)
+	ListUsers(ctx context.Context, page, limit int) ([]domain.User, error)
 	UpdateUser(ctx context.Context, user *domain.User) (*domain.User, error)
 	DeleteUser(ctx context.Context, id string) error
 }
